Extract rabbitMQ from `matchingController.js`
